'DiskEncryption' --> 'Encryption' for folder and component, updated import paths accordingly; slight refactors in Encryption.tsx to make it more flexible for upcoming Volumes use case
