+++ packages/manager/src/components/Encryption/Encryption.tsx
@@ -3,35 +3,38 @@ import * as React from 'react';
 import { Box } from 'src/components/Box';
 import { Checkbox } from 'src/components/Checkbox';
 import { Typography } from 'src/components/Typography';
+
 import { Notice } from '../Notice/Notice';
 
-export interface DiskEncryptionProps {
+export interface EncryptionProps {
   descriptionCopy: JSX.Element | string;
   disabled?: boolean;
   disabledReason?: string;
+  entityType?: string;
   error?: string;
-  isEncryptDiskChecked: boolean;
+  isEncryptEntityChecked: boolean;
   onChange: (checked: boolean) => void;
 }
 
-export const headerTestId = 'disk-encryption-header';
-export const descriptionTestId = 'disk-encryption-description';
-export const checkboxTestId = 'encrypt-disk-checkbox';
+export const headerTestId = 'encryption-header';
+export const descriptionTestId = 'encryption-description';
+export const checkboxTestId = 'encrypt-entity-checkbox';
 
-export const DiskEncryption = (props: DiskEncryptionProps) => {
+export const Encryption = (props: EncryptionProps) => {
   const {
     descriptionCopy,
     disabled,
     disabledReason,
+    entityType,
     error,
-    isEncryptDiskChecked,
+    isEncryptEntityChecked,
     onChange,
   } = props;
 
   return (
     <>
       <Typography data-testid={headerTestId} variant="h3">
-        Disk Encryption
+        {`${entityType ?? 'Disk'} Encryption`}
       </Typography>
       {error && (
         <Notice spacingBottom={0} spacingTop={8} text={error} variant="error" />
@@ -51,11 +54,11 @@ export const DiskEncryption = (props: DiskEncryptionProps) => {
         flexDirection="row"
       >
         <Checkbox
-          checked={disabled ? false : isEncryptDiskChecked} // in Create flows, this will be defaulted to be checked. Otherwise, we will rely on the current encryption status for the initial value
+          checked={disabled ? false : isEncryptEntityChecked} // in Create flows, this will be defaulted to be checked. Otherwise, we will rely on the current encryption status for the initial value
           data-testid={checkboxTestId}
           disabled={disabled}
           onChange={(e, checked) => onChange(checked)}
-          text="Encrypt Disk"
+          text={`Encrypt ${entityType ?? 'Disk'}`}
           toolTipText={disabled ? disabledReason : ''}
         />
       </Box>
