+++ packages/manager/src/components/Encryption/utils.ts
