+++ packages/manager/src/features/Linodes/LinodeCreatev2/Region.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import { useController, useFormContext, useWatch } from 'react-hook-form';
 
 import { Box } from 'src/components/Box';
-import { useIsDiskEncryptionFeatureEnabled } from 'src/components/DiskEncryption/utils';
 import { DocsLink } from 'src/components/DocsLink/DocsLink';
+import { useIsDiskEncryptionFeatureEnabled } from 'src/components/Encryption/utils';
 import { Link } from 'src/components/Link';
 import { Notice } from 'src/components/Notice/Notice';
 import { Paper } from 'src/components/Paper';
