+++ packages/manager/src/features/Linodes/LinodeEntityDetail.test.tsx
@@ -42,9 +42,9 @@ describe('Linode Entity Detail', () => {
     };
   });
 
-  vi.mock('src/components/DiskEncryption/utils.ts', async () => {
+  vi.mock('src/components/Encryption/utils.ts', async () => {
     const actual = await vi.importActual<any>(
-      'src/components/DiskEncryption/utils.ts'
+      'src/components/Encryption/utils.ts'
     );
     return {
       ...actual,
